Implement factory function for event and control datagram types

+++ ground_station/3D/Objects/Ruler.cpp
@@ -53,7 +53,8 @@ void Ruler::initLabels(Qt3DRender::QCamera *camera, const QString &tickTexture) 
          i <= OpenGLConstants::RULER_MAX_HEIGHT;
          i += OpenGLConstants::RULER_SPACING) {
         tickLabels_.push_back(
-                new Tracker({0, i, 0}, camera, tickTexture, QString::number(i) + QStringLiteral(" m"), TextType::BOLD,
+                new Tracker({0, static_cast<float>(i), 0}, camera, tickTexture,
+                            QString::number(i) + QStringLiteral(" m"), TextType::BOLD,
                             this,
                             OpenGLConstants::RIGHT_TICK, OpenGLConstants::RIGHT_LABEL)
         );
